fix defects
