+++ Source/TheAfterlife/HomeTheAfterlife/Core/Characters/Animations/Notifies/AnimNotify_EnableRagdoll.cpp
@@ -1,8 +1,9 @@
 
 #include "AnimNotify_EnableRagdoll.h"
+#include "../../../../../TheAfterlifeTypes.h"
 
 void UAnimNotify_EnableRagdoll::Notify(USkeletalMeshComponent* MeshComp, UAnimSequenceBase* Animation)
 {
-	MeshComp->SetCollisionProfileName(FName("Ragdoll"));
+	MeshComp->SetCollisionProfileName(CollisionProfileRagdoll);
 	MeshComp->SetSimulatePhysics(true);
 }
