+++ Source/TheAfterlife/HomeTheAfterlife/Core/Characters/PlayerCharacter.h
@@ -25,8 +25,6 @@ public:
 	virtual void LookUp(float value) override;
 	virtual void Jump() override;
 	virtual void Landed(const FHitResult& Hit) override;
-	virtual void Falling() override;
-	virtual void NotifyJumpApex() override;
 
 protected:
 
@@ -39,21 +37,6 @@ protected:
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Character|Jump")
 	class UAnimMontage* DoubleJumpMontage;
 
-	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = "Character|Components")
-	class UCharacterAttributeComponent* CharacterAttributeComponent;
-
-	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Character | Animations")
-	class UAnimMontage* OnDeathAnimMontage;
-	// Damage depending from fall height (in meters)
-	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Character | Attributes")
-	class UCurveFloat* FallDamageCurve;
-
-	virtual void OnDeath();
-
 private:
 	int32 JumpCount = 0;
-	
-	void EnableRagdoll();
-
-	FVector CurrentFallApex;
 };
